+++ src/services/Sso.php
@@ -217,26 +217,18 @@ class Sso extends Component
      */
     public function findUser(ProviderInterface $provider, string $idpIdentifier): ?User
     {
-        $userId = (new Query())
-            ->select([
-                'userId',
-            ])
-            ->from([
-                Table::SSO_IDENTITIES,
-            ])
-            ->where(
+        return User::find()
+            ->innerJoin(
+                ['s_i' => Table::SSO_IDENTITIES],
+                '[[s_i.userId]] = [[users.id]]',
+            )
+            ->andWhere(
                 [
-                    'provider' => $provider->getHandle(),
-                    'identityId' => $idpIdentifier,
+                    's_i.provider' => $provider->getHandle(),
+                    's_i.identityId' => $idpIdentifier,
                 ]
             )
-            ->scalar();
-
-        if (!$userId) {
-            return null;
-        }
-
-        return Craft::$app->getUsers()->getUserById($userId);
+            ->one();
     }
 
     /**